requirements.txt
# Core ML Dependencies
torch>=2.4.0  # PyTorch 2.5.0+ recommended for enable_gqa support
accelerate>=0.24.0
transformers>=4.49.0
diffusers>=0.30.0

# Supporting dependencies
protobuf
numpy
tqdm
requests
matplotlib
pillow
ipython
seaborn
sentencepiece
open-clip-torch  # For LAION aesthetic predictor

# VLM Quality Assessment Dependencies
qwen-vl-utils[decord]
flash-attn>=2.0.0
einops
timm
